+++ assets/system/cell_modules/cell-runtime.js
@@ -1,0 +1,19 @@
+/***
+ *   Sphere v2 Cell Runtime master module
+  *  (c) 2017 Fat Cerberus
+***/
+
+'use strict';
+exports.__esModule = true;
+
+Object.assign(exports,
+{
+	assert:     require('assert'),
+	from:       require('from'),
+	transpile:  require('transpile'),
+	DataReader: require('dataReader'),
+	DataWriter: require('dataWriter'),
+	Random:     require('random'),
+	Test:       require('test'),
+	XML:        require('xml'),
+});