+++ assets/system/cell_modules/dataReader.js
@@ -1,21 +1,14 @@
 /**
- *  miniRT struct CommonJS module
- *  (c) 2016 Fat Cerberus
+ *  Sphere v2 Runtime: DataReader object
+ *  (c) 2015-2017 Fat Cerberus
 **/
 
 'use strict';
-module.exports =
-{
-	Reader: Reader,
-	Writer: Writer,
-};
-
 const assert = require('assert');
-const from   = require('from');
 
-function Reader(stream)
+function DataReader(stream)
 {
-	assert(this instanceof Reader, "constructor requires 'new'");
+	assert(this instanceof DataReader, "constructor requires 'new'");
 	assert('read' in stream, "not a readable stream");
 
 	var m_decoder = new TextDecoder('utf-8');
@@ -156,174 +149,6 @@ function Reader(stream)
 		assert.equal(typeof littleEndian, 'boolean');
 
 		return _readInt(m_stream, 4, false, littleEndian);
-	}
-}
-
-function Writer(stream)
-{
-	assert(this instanceof Writer, "constructor requires 'new'");
-	assert('write' in stream, "not a writable stream");
-
-	var m_encoder = new TextEncoder();
-	var m_stream = stream;
-
-	this.writeFloat32 = m_writeFloat32;
-	function m_writeFloat32(value, littleEndian)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(4));
-		dv.setFloat32(0, value, littleEndian);
-		m_stream.write(dv);
-	}
-
-	this.writeFloat64 = m_writeFloat64;
-	function m_writeFloat64(value, littleEndian)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(8));
-		dv.setFloat64(0, value, littleEndian);
-		m_stream.write(dv);
-	}
-
-	this.writeInt8 = m_writeInt8;
-	function m_writeInt8(value)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(1));
-		dv.setInt8(0, value);
-		m_stream.write(dv);
-	}
-
-	this.writeInt16 = m_writeInt16;
-	function m_writeInt16(value, littleEndian)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(2));
-		dv.setInt16(0, value, littleEndian);
-		m_stream.write(dv);
-	}
-
-	this.writeInt32 = m_writeInt32;
-	function m_writeInt32(value, littleEndian)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(4));
-		dv.setInt32(0, value, littleEndian);
-		m_stream.write(dv);
-	}
-
-	this.writeString = m_writeString;
-	function m_writeString(value, length)
-	{
-		assert.equal(typeof value, 'string');
-		assert.equal(typeof length, 'number');
-
-		var encoded = m_encoder.encode(value);
-		var bytes = new Uint8Array(length);
-		bytes.set(encoded.subarray(0, length));
-		m_stream.write(bytes);
-	}
-
-	this.writeString8 = m_writeString8;
-	function m_writeString8(value)
-	{
-		assert.equal(typeof value, 'string');
-
-		var bytes = m_encoder.encode(value);
-		m_writeUint8(bytes.length);
-		m_stream.write(bytes);
-	}
-
-	this.writeString16 = m_writeString16;
-	function m_writeString16(value, littleEndian)
-	{
-		assert.equal(typeof value, 'string');
-
-		var bytes = m_encoder.encode(value);
-		m_writeUint16(bytes.length, littleEndian);
-		m_stream.write(bytes);
-	}
-
-	this.writeString32 = m_writeString32;
-	function m_writeString32(value, littleEndian)
-	{
-		assert.equal(typeof value, 'string');
-
-		var bytes = m_encoder.encode(value);
-		m_writeUint32(bytes.length, littleEndian);
-		m_stream.write(bytes);
-	}
-
-	this.writeStruct = m_writeStruct;
-	function m_writeStruct(object, desc)
-	{
-		_checkStructDescriptor(desc);
-
-		var keys = Reflect.ownKeys(desc);
-		for (var i = 0; i < keys.length; ++i) {
-			var key = keys[i];
-			var value = key in object ? object[key]
-				: desc[key].default
-			switch (desc[key].type) {
-				case 'bool': m_writeUint8(value ? 1 : 0); break;
-				case 'float32be': m_writeFloat32(value); break;
-				case 'float32le': m_writeFloat32(value, true); break;
-				case 'float64be': m_writeFloat64(value); break;
-				case 'float64le': m_writeFloat64(value, true); break;
-				case 'int8': m_writeInt8(value); break;
-				case 'int16be': m_writeInt16(value); break;
-				case 'int16le': m_writeInt16(value, true); break;
-				case 'int32be': m_writeInt32(value); break;
-				case 'int32le': m_writeInt32(value, true); break;
-				case 'uint8': m_writeUint8(value); break;
-				case 'uint16be': m_writeUint16(value); break;
-				case 'uint16le': m_writeUint16(value, true); break;
-				case 'uint32be': m_writeUint32(value); break;
-				case 'uint32le': m_writeUint32(value, true); break;
-				case 'fstring': m_writeString(value, desc[key].length); break;
-				case 'lstr8': m_writeString8(value); break;
-				case 'lstr16be': m_writeString16(value); break;
-				case 'lstr16le': m_writeString16(value, true); break;
-				case 'lstr32be': m_writeString32(value); break;
-				case 'lstr32le': m_writeString32(value, true); break;
-				case 'raw': m_stream.write(value); break;
-			}
-		}
-	};
-
-	this.writeUint8 = m_writeUint8;
-	function m_writeUint8(value)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(1));
-		dv.setUint8(0, value);
-		m_stream.write(dv);
-	}
-
-	this.writeUint16 = m_writeUint16;
-	function m_writeUint16(value, littleEndian)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(2));
-		dv.setUint16(0, value, littleEndian);
-		m_stream.write(dv);
-	}
-
-	this.writeUint32 = m_writeUint32;
-	function m_writeUint32(value, littleEndian)
-	{
-		assert.equal(typeof value, 'number');
-
-		var dv = new DataView(new ArrayBuffer(4));
-		dv.setUint32(0, value, littleEndian);
-		m_stream.write(dv);
 	}
 }