+++ assets/system/cell_modules/from.js
@@ -4,10 +4,17 @@
 **/
 
 'use strict';
-module.exports = from;
+exports = module.exports = from;
+exports.__esModule = true;
+exports.default = exports;
 
-const assert = require('assert');
-const random = require('random');
+const assert = require('assert'),
+      Random = require('random');
+
+const PK =
+{
+	ItemSource: Symbol('itemSource'),
+};
 
 function from(target/*, ...*/)
 {
@@ -51,11 +58,6 @@ function fromIterable(target)
 	var itemSource = new IterableSource(target);
 	return new FromQuery(itemSource);
 }
-
-const PK =
-{
-	ItemSource: Symbol('itemSource'),
-};
 
 function MAKEPOINT(sourceType, op)
 {
@@ -507,7 +509,7 @@ function SampleSource(uniqueOnly)
 			var item;
 
 			if (m_numSamples++ < m_count) {
-				index = random.discrete(0, m_items.length - 1);
+				index = Random.discrete(0, m_items.length - 1);
 				item = m_items[index];
 				if (uniqueOnly)
 					m_items.splice(index, 1);
@@ -540,7 +542,7 @@ function ShuffledSource(source)
 			m_list[index] = item;
 		}
 		for (var i = m_list.length - 1; i >= 1; --i) {
-			index = random.discrete(0, i);
+			index = Random.discrete(0, i);
 			temp = m_list[index];
 			m_list[index] = m_list[i];
 			m_list[i] = temp;