+++ assets/system/cell_modules/xml.js
@@ -4,20 +4,19 @@
 **/
 
 'use strict';
-module.exports =
-{
-	load:  load,
-	parse: parse,
-};
+exports.__esModule = true;
+exports.default = exports;
 
 const sax = require('./lib/sax');
 
-function load(fileName)
+exports.readFile = readFile;
+function readFile(fileName)
 {
 	var xmlText = FS.readFile(fileName);
 	return parse(xmlText);
 }
 
+exports.parse = parse;
 function parse(xmlText)
 {
 	var dom = { type: 'root', nodes: [] };