+++ CHANGELOG.md
@@ -4,12 +4,18 @@ miniSphere Changelog
 vX.X.X - TBD
 ------------
 
+* Adds a new `cell-runtime` master module for use in ES6 Cellscripts, as an
+  analogue to `sphere-runtime`.
 * Renames `FS.exists()` to `FS.fileExists()` since the function doesn't work
   for directories.
 * Renames `FS.resolve()` to `FS.fullPath()` to improve accessibility.
 * Renames `FileStream#size` to `FileStream#fileSize`.
+* Fixes some inconsistencies between modules in the Cell Runtime and their
+  Sphere Runtime equivalents.
 * Fixes an issue where chaining `Dispatch.now()` causes the engine to get stuck
   in an infinite loop of processing newly added jobs.
+* Fixes a bug where `DataReader` and `DataWriter` can't be imported by either
+  `require()` or `import`.
 * Fixes a bug where constructors can be called without using `new`.
 * Fixes a bug where stepping over a `Dispatch.now()` with SSj sometimes locks
   up the engine and prevents it from responding to further debugger commands.